Update more room components and less zagjs!
