--- web/src/pages/Room/components/VoterTable/components/OptionConfirmationDialog/OptionConfirmationDialog.module.scss
@@ -1,52 +1,0 @@
-@use "../../../../../../index.scss";
-
-.optionConfirmationBackdrop {
-	background-color: rgba(0, 0, 0, 0.25);
-	position: absolute;
-	inset: 0;
-}
-
-.optionConfirmationContainer {
-	animation: flipIn 200ms cubic-bezier(0.165, 0.84, 0.44, 1);
-
-	background-color: var(--elevated-background);
-	border-radius: 0.5rem;
-	padding: 1rem 1.25rem;
-	box-shadow: var(--shadow-xl);
-
-	position: absolute;
-	inset: 0;
-	width: 25rem;
-	max-width: 90%;
-	height: fit-content;
-	margin: auto;
-
-	[data-part="close-trigger"] {
-		position: absolute;
-		color: var(--primary-text-color);
-		top: 0.25rem;
-		right: 0.25rem;
-		width: 2rem;
-		height: 2rem;
-		padding: 0;
-	}
-
-	[data-part="title"] {
-		font-size: 1.25rem;
-	}
-
-	[data-part="description"] {
-		padding-top: 1.25rem;
-		padding-bottom: 2rem;
-	}
-
-	div[role="group"] {
-		display: flex;
-		gap: 0.5rem;
-		justify-content: flex-end;
-
-		& > button {
-			padding: 0.3rem 0.5rem;
-		}
-	}
-}
