--- web/src/pages/Room/components/VoterTable/components/OptionConfirmationDialog/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./OptionConfirmationDialog";
