--- web/src/pages/Room/components/VoterTable/components/VoterOptionsMenu/VoterOptionsMenu.module.scss
@@ -1,20 +1,0 @@
-.voterActionMenu {
-	background-color: var(--elevated-background);
-	border-radius: 0.5rem;
-	overflow: hidden;
-	box-shadow: var(--shadow-xl);
-
-	.itemIcon {
-		margin-right: 0.25rem;
-	}
-
-	[data-part="item"] {
-		list-style-type: none;
-		padding: 0.5rem 0.75rem;
-		cursor: pointer;
-
-		&[data-focus] {
-			background-color: var(--alpha-orange);
-		}
-	}
-}
