--- web/src/pages/Room/components/VoterTable/components/VoterOptionsMenu/index.ts
@@ -1,1 +1,0 @@
-export { default, type VoterClickAction } from "./VoterOptionsMenu";
