--- web/src/pages/Room/views/moderator/ModeratorView.module.scss
@@ -1,4 +1,0 @@
-button.moderatorBtn {
-	width: 100%;
-	margin-bottom: 1rem;
-}
