--- web/src/pages/Room/views/moderator/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./ModeratorView";
