--- web/src/pages/Room/views/voter/VoterView.module.scss
@@ -1,25 +1,0 @@
-fieldset.voterChoices {
-	border: none;
-	margin: 0 auto;
-	padding: 0;
-	display: flex;
-	flex-wrap: wrap;
-	gap: 0.5rem;
-	justify-content: center;
-
-	legend {
-		font-size: 1.1rem;
-		font-weight: 600;
-		margin-bottom: 1rem;
-		text-align: center;
-		width: 100%;
-	}
-}
-
-p.infoText {
-	font-weight: 600;
-	margin-bottom: 1rem;
-	white-space: nowrap;
-	overflow: hidden;
-	text-overflow: ellipsis;
-}
