--- web/src/pages/Room/views/voter/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./VoterView";
